frontend: Rename Instance/StatusHistoryItem.react.js to tsx
